requirements.txt
datasets
transformers
tokenizers
langchain
langchain-community
tiktoken
modelscope
sentence-transformers
chromadb
gradio
deepspeed
duckduckgo_search
open-retrievals
rank_bm25

fastapi
unicorn
pydantic_settings
lmdeploy
